TES-51 feat: improve zod validation
