+++ src/components/PlaceOrderModal.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import React from 'react';
-import { Button } from '@/components/ui/button';
 import { OrderNowBtn } from '@/components/OrderNowBtn';
 import {
   Dialog,
@@ -11,39 +10,14 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useForm } from 'react-hook-form';
-import { z } from 'zod';
+import { PlaceOrderForm } from '@/components/forms/PlaceOrderForm';
 
 interface PlaceOrderModalProps {
   isOpen: boolean;
   setIsOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-const FormSchema = z.object({
-  nameAndSurname: z.string(),
-  email: z.string(),
-  phoneNumber: z.string(),
-});
-
 export const PlaceOrderModal: React.FC<PlaceOrderModalProps> = () => {
-  const form = useForm<z.infer<typeof FormSchema>>({
-    resolver: zodResolver(FormSchema),
-  });
-
-  const onSubmit = (data: z.infer<typeof FormSchema>) => {
-    console.log(data);
-  };
-
   return (
     <Dialog>
       <DialogTrigger asChild>
@@ -68,58 +42,7 @@ export const PlaceOrderModal: React.FC<PlaceOrderModalProps> = () => {
               very shortly to finalise your order.
             </DialogDescription>
           </DialogHeader>
-          <Form {...form}>
-            <form
-              onSubmit={form.handleSubmit(onSubmit)}
-              className='mt-4 flex flex-col gap-[20px]'
-            >
-              <FormField
-                control={form.control}
-                name='nameAndSurname'
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>* Name and surname</FormLabel>
-                    <FormControl>
-                      <Input {...field} className='text-grey-400' />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name='email'
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>* Email address</FormLabel>
-                    <FormControl>
-                      <Input {...field} className='text-grey-400' />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name='phoneNumber'
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>* Phone number</FormLabel>
-                    <FormControl>
-                      <Input {...field} className='text-grey-400' />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <Button
-                type='submit'
-                className='bg-primary-dark mt-3 max-w-[156px] md:mt-7'
-              >
-                Place order
-              </Button>
-            </form>
-          </Form>
+          <PlaceOrderForm />
         </div>
       </DialogContent>
     </Dialog>
